+++ dart/geobase/lib/src/vector/formats/wkt/wkt_like_format.dart
@@ -34,6 +34,13 @@ import '/src/vector/encoding/text_format.dart';
 /// bounding boxes are formatted as polygons. An example presented above however
 /// format bounding box as a point series of two points (min, max). See also
 /// `wktFormat` that formats them as polygons.
+/// 
+/// When getting an encoder from text writer format objects this `WktLikeFormat`
+/// class provides you COULD use `crs` parameter to give hints (like axis order,
+/// and whether x and y must be swapped when writing) about coordinate reference
+/// system in text output. However for the WKT (like) text format such crs
+/// information is ignored, x/longitude is always printed before y/latitude
+/// regardless of crs axis order.
 class WktLikeFormat {
   /// The WKT (like) text writer format for coordinate objects.
   static const TextWriterFormat<CoordinateContent> coordinate =