+++ dart/geobase/lib/src/vector_data/model/geometry/geometry.dart
@@ -10,6 +10,7 @@ import 'package:meta/meta.dart';
 
 import '/src/codes/coords.dart';
 import '/src/codes/geom.dart';
+import '/src/coordinates/crs/coord_ref_sys.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/vector/content/geometry_content.dart';
 import '/src/vector/content/simple_geometry_content.dart';
@@ -49,14 +50,19 @@ abstract class Geometry extends Bounded {
   ///
   /// Use [decimals] to set a number of decimals (not applied if no decimals).
   ///
+  /// Use [crs] to give hints (like axis order, and whether x and y must
+  /// be swapped when writing) about coordinate reference system in text output.
+  ///
   /// Other format or encoder implementation specific options can be set by
   /// [options].
   String toText({
     TextWriterFormat<GeometryContent> format = GeoJSON.geometry,
     int? decimals,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) {
-    final encoder = format.encoder(decimals: decimals, options: options);
+    final encoder =
+        format.encoder(decimals: decimals, crs: crs, options: options);
     writeTo(encoder.writer);
     return encoder.toText();
   }
@@ -104,9 +110,11 @@ abstract class SimpleGeometry extends Geometry {
   String toText({
     TextWriterFormat<SimpleGeometryContent> format = GeoJSON.geometry,
     int? decimals,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) {
-    final encoder = format.encoder(decimals: decimals, options: options);
+    final encoder =
+        format.encoder(decimals: decimals, crs: crs, options: options);
     writeTo(encoder.writer);
     return encoder.toText();
   }